Restructure identity client tests
- Moves federation tests into its own file
- Adds test helpers file to be used across test files
- Updating internal login lock tests that are failing
